requirements.txt
asgiref==3.7.2
Django==4.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.6.3
whitenoise==6.5.0
